이미지 렌더링

+++ Engine/Source/Application/WindowsLaunch/Private/WindowsErrors.inl
@@ -25,5 +25,5 @@ inline void ReportWindowsError(std::wstring_view InMessage = L"", int32 Error = 
 inline void ReportCOMError(std::wstring_view InMessage, HRESULT Error)
 {
 	_com_error ComError(Error);
-	SE_LOG(LogWindowsLaunch, Fatal, L"An error occurred from COM: {}\nErrorCode: 0x{:08X}, FormattedMessage: {}", InMessage, Error, TCHAR_TO_WCHAR(ComError.ErrorMessage()));
+	SE_LOG(LogWindowsLaunch, Fatal, L"An error occurred from COM: {}\nErrorCode: 0x{:08X}, FormattedMessage: {}", InMessage, (uint32)Error, TCHAR_TO_WCHAR(ComError.ErrorMessage()));
 }
